# Backend/swagger_codegen/swagger_server/models/get_pollution_info_data.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class GetPollutionInfoData(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, lat: float=None, lng: float=None, range: float=None):  # noqa: E501
        """GetPollutionInfoData - a model defined in Swagger

        :param lat: The lat of this GetPollutionInfoData.  # noqa: E501
        :type lat: float
        :param lng: The lng of this GetPollutionInfoData.  # noqa: E501
        :type lng: float
        :param range: The range of this GetPollutionInfoData.  # noqa: E501
        :type range: float
        """
        self.swagger_types = {
            'lat': float,
            'lng': float,
            'range': float
        }

        self.attribute_map = {
            'lat': 'lat',
            'lng': 'lng',
            'range': 'range'
        }
        self._lat = lat
        self._lng = lng
        self._range = range

    @classmethod
    def from_dict(cls, dikt) -> 'GetPollutionInfoData':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The GetPollutionInfoData of this GetPollutionInfoData.  # noqa: E501
        :rtype: GetPollutionInfoData
        """
        return util.deserialize_model(dikt, cls)

    @property
    def lat(self) -> float:
        """Gets the lat of this GetPollutionInfoData.

        Latitude  # noqa: E501

        :return: The lat of this GetPollutionInfoData.
        :rtype: float
        """
        return self._lat

    @lat.setter
    def lat(self, lat: float):
        """Sets the lat of this GetPollutionInfoData.

        Latitude  # noqa: E501

        :param lat: The lat of this GetPollutionInfoData.
        :type lat: float
        """

        self._lat = lat

    @property
    def lng(self) -> float:
        """Gets the lng of this GetPollutionInfoData.

        Longitude  # noqa: E501

        :return: The lng of this GetPollutionInfoData.
        :rtype: float
        """
        return self._lng

    @lng.setter
    def lng(self, lng: float):
        """Sets the lng of this GetPollutionInfoData.

        Longitude  # noqa: E501

        :param lng: The lng of this GetPollutionInfoData.
        :type lng: float
        """

        self._lng = lng

    @property
    def range(self) -> float:
        """Gets the range of this GetPollutionInfoData.

        Range in meters to get sensor data  # noqa: E501

        :return: The range of this GetPollutionInfoData.
        :rtype: float
        """
        return self._range

    @range.setter
    def range(self, range: float):
        """Sets the range of this GetPollutionInfoData.

        Range in meters to get sensor data  # noqa: E501

        :param range: The range of this GetPollutionInfoData.
        :type range: float
        """

        self._range = range