# Backend/swagger_codegen/swagger_server/models/pollution_info.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class PollutionInfo(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, lat: float=None, lng: float=None):  # noqa: E501
        """PollutionInfo - a model defined in Swagger

        :param lat: The lat of this PollutionInfo.  # noqa: E501
        :type lat: float
        :param lng: The lng of this PollutionInfo.  # noqa: E501
        :type lng: float
        """
        self.swagger_types = {
            'lat': float,
            'lng': float
        }

        self.attribute_map = {
            'lat': 'lat',
            'lng': 'lng'
        }
        self._lat = lat
        self._lng = lng

    @classmethod
    def from_dict(cls, dikt) -> 'PollutionInfo':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The PollutionInfo of this PollutionInfo.  # noqa: E501
        :rtype: PollutionInfo
        """
        return util.deserialize_model(dikt, cls)

    @property
    def lat(self) -> float:
        """Gets the lat of this PollutionInfo.

        Latitude  # noqa: E501

        :return: The lat of this PollutionInfo.
        :rtype: float
        """
        return self._lat

    @lat.setter
    def lat(self, lat: float):
        """Sets the lat of this PollutionInfo.

        Latitude  # noqa: E501

        :param lat: The lat of this PollutionInfo.
        :type lat: float
        """

        self._lat = lat

    @property
    def lng(self) -> float:
        """Gets the lng of this PollutionInfo.

        Longitude  # noqa: E501

        :return: The lng of this PollutionInfo.
        :rtype: float
        """
        return self._lng

    @lng.setter
    def lng(self, lng: float):
        """Sets the lng of this PollutionInfo.

        Longitude  # noqa: E501

        :param lng: The lng of this PollutionInfo.
        :type lng: float
        """

        self._lng = lng